Moved code setting initial value to observe method
Before the change, the initial value of the "Bindable" object is set
only by the Bindable.bind() method.  This change moved that initial set
value code to the NSObject.observe() method in the NSObject extension.
The Bindable.bind() method calls the NSObject.observe() method within
it.  As a result, the initial value is set regardless of which method is
called.